refactored main, split configs
This is a very clarifying commit on the boundaries. The main file was
getting too big as well.

I split it into 3 where 2 I am pretty sure about:

- network
- image
- sync

Network has just the peers basically and the configuration about how
they attach (mount) to each other. It is here that you can get
contexts/connections to different peers in the... network.

The image is the thing that includes the replicas and the paths of
those replicas on the peers.

This is where the working sets, replicas, and overall image thing
lives.
The key is that the Image has to be associated with a network object
so we can get the concrete connections relative to ourselves.

I kind of pushed the other stuff into sync. It will likely get split
up itself as it becomes more clear about how to implement that.